fix: `exit 1` properly exits
In this case, there is no difference because the the script currently exists with the exit code of the subshell - 1.

But later, code may be added - in which case, the `exit 1` has no effect (because it is in a subshell). And `set -e` isn't set for the script as a whole.